requirements.txt
streamlit>=1.28.0
pillow>=9.0.0
pygame>=2.0.0
pygame-gui>=0.6.0 